+++ roulier/carriers/laposte/laposte_encoder.py
@@ -37,7 +37,7 @@ class LaposteEncoder(Encoder):
         return {
             "body": template.render(
                 service=data['service'],
-                parcel=data['parcel'],
+                parcel=data['parcels'][0],
                 auth=data['auth'],
                 sender_address=data['from_address'],
                 receiver_address=data['to_address'],